Some new changes.

+++ 07_DAO_Spring_JDBC/src/main/resources/schema.sql
@@ -1,22 +1,22 @@
 DROP TABLE IF EXISTS author CASCADE;
 CREATE TABLE author(
-    id INT,
+    id BIGSERIAL,
     name CHARACTER VARYING(255),
     CONSTRAINT author_pk PRIMARY KEY (id)
 );
 
 DROP TABLE IF EXISTS genre CASCADE;
 CREATE TABLE genre(
-    id INT,
+    id BIGSERIAL,
     name CHARACTER VARYING(50),
     CONSTRAINT genre_pk PRIMARY KEY (id)
 );
 
 DROP TABLE IF EXISTS book;
 CREATE TABLE book(
-    id int,
-    author_id int NOT NULL,
-    genre_id int NOT NULL,
+    id BIGSERIAL,
+    author_id BIGINT NOT NULL,
+    genre_id BIGINT NOT NULL,
     name CHARACTER VARYING(50),
     CONSTRAINT book_pk PRIMARY KEY (id)
 );
